Remove problem with nextInt() passing a newline on to the Scanner
Turns out the problem was in the nextInt() method in Scanner class, which only reads the integers in the input, leaving the newline used to enter as input for the following nextLine() method.

This was solved using Integer.parseInt(scanner.newLine()) instead of newInt().

NEXT STEP:

Find a way to prevent user from entering unwanted values into the Scanner.