Add methods for handling JSON, resolve contactID problems
Created system.JSONController class which contains methods for:
        - turning Contact objects into JSONObjects,
        - turning JSONObjects into Contact objects,
        - turning a Contact[] into a JSONArray consisiting of
          Contact objects turned into JSONObjects
        - writing a JSONObject or JSONArray into a file
	  (contactBookData.json)

Resolved problems with the numberOfContacts field in Contact
class. It is now split into a nextID field and a separate
counter in class ContactBook which is incremented or decre-
mented when adding Contacts to the ContactBook using its
internal methods.

NEXT STEP:
	Create method for loading a .json file and
	turning it into a Contact[] which can then
	be loaded into a ContactBook instance.